requirements.txt
pypdf2
pymupdf
python-dotenv
openai
reportlab